Rename `actionButtons.tsx` to `ActionButtons.tsx`
